camera_simulator/requirements.txt
flask
opencv-python
requests
python-dotenv
flask-cors
